feat: add temp blob storage

+++ src/Arcus.Testing.Storage.Blob/TemporaryBlobFile.cs
@@ -2,40 +2,81 @@
 using System.Threading.Tasks;
 using Azure.Identity;
 using Azure.Storage.Blobs;
+using Azure.Storage.Blobs.Models;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 
 namespace Arcus.Testing
 {
     /// <summary>
+    /// Represents the available options when creating a <see cref="TemporaryBlobFile"/>.
+    /// </summary>
+    public class OnSetupBlobFile
+    {
+        internal bool Override { get; private set; }
+
+        /// <summary>
+        /// (default) Configures the <see cref="TemporaryBlobFile"/> to override an existing Azure Blob file when it already exists.
+        /// </summary>
+        public OnSetupBlobFile OverrideExistingBlob()
+        {
+            Override = true;
+            return this;
+        }
+
+        /// <summary>
+        /// Configures the <see cref="TemporaryBlobFile"/> to use the existing Azure Blob file's content instead when it already exists.
+        /// </summary>
+        public OnSetupBlobFile UseExistingBlob()
+        {
+            Override = false;
+            return this;
+        }
+    }
+
+    internal enum UponBlobDelete { DeleteIfCreated, DeleteIfExisted }
+
+    /// <summary>
+    /// Represents the available options when deleting a <see cref="TemporaryBlobFile"/>.
+    /// </summary>
+    public class OnTeardownBlobFile
+    {
+        internal UponBlobDelete Content { get; private set; }
+
+        /// <summary>
+        /// (default) Configures the <see cref="TemporaryBlobFile"/> to delete the Azure Blob file upon disposal if the test fixture created the file.
+        /// </summary>
+        public OnTeardownBlobFile DeleteCreatedBlob()
+        {
+            Content = UponBlobDelete.DeleteIfCreated;
+            return this;
+        }
+
+        /// <summary>
+        /// Configures the <see cref="TemporaryBlobFile"/> to delete the Azure Blob file upon disposal, even if it already existed - outside the fixture's scope.
+        /// </summary>
+        /// <returns></returns>
+        public OnTeardownBlobFile DeleteExistingBlob()
+        {
+            Content = UponBlobDelete.DeleteIfExisted;
+            return this;
+        }
+    }
+
+    /// <summary>
     /// Represents the available options when uploading a <see cref="TemporaryBlobFile"/>.
     /// </summary>
     public class TemporaryBlobFileOptions
     {
-        private string _blobName = $"test-{Guid.NewGuid()}";
+        /// <summary>
+        /// Gets the additional options to manipulate the creation of the <see cref="TemporaryBlobFile"/>.
+        /// </summary>
+        public OnSetupBlobFile OnSetup { get; } = new OnSetupBlobFile().UseExistingBlob();
 
         /// <summary>
-        /// Gets or sets the name of the Azure Blob file to upload.
+        /// Gets the additional options to manipulate the deletion of the <see cref="TemporaryBlobFile"/>.
         /// </summary>
-        /// <exception cref="ArgumentException">Thrown when the <paramref name="value"/> is blank.</exception>
-        public string BlobName
-        {
-            get => _blobName;
-            set
-            {
-                if (string.IsNullOrWhiteSpace(value))
-                {
-                    throw new ArgumentException("Requires a non-blank Azure Blob name to upload a blob to the storage", nameof(value));
-                }
-
-                _blobName = value;
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the flag indicating whether or not to override an existing blob with the same name (default: <c>true</c>).
-        /// </summary>
-        public bool OverrideExistingBlob { get; set; } = true;
+        public OnTeardownBlobFile OnTeardown { get; } = new OnTeardownBlobFile().DeleteCreatedBlob();
     }
 
     /// <summary>
@@ -44,18 +85,34 @@ namespace Arcus.Testing
     public class TemporaryBlobFile : IAsyncDisposable
     {
         private readonly BlobClient _blobClient;
+        private readonly bool _createdByUs;
+        private readonly BinaryData _originalData;
+        private readonly TemporaryBlobFileOptions _options;
         private readonly ILogger _logger;
 
-        private TemporaryBlobFile(BlobClient blobClient, ILogger logger)
+        private TemporaryBlobFile(
+            BlobClient blobClient,
+            bool createdByUs,
+            BinaryData originalData,
+            TemporaryBlobFileOptions options,
+            ILogger logger)
         {
             _blobClient = blobClient ?? throw new ArgumentNullException(nameof(blobClient));
+            _createdByUs = createdByUs;
+            _originalData = originalData;
+            _options = options;
             _logger = logger ?? NullLogger.Instance;
         }
 
         /// <summary>
         /// Gets the name of the Azure Blob file currently in storage.
         /// </summary>
-        public string FileName => _blobClient.Name;
+        public string Name => _blobClient.Name;
+
+        /// <summary>
+        /// Gets the client to interact with the temporary stored Azure Blob file currently in storage.
+        /// </summary>
+        public BlobClient Client => _blobClient;
 
         /// <summary>
         /// Uploads a temporary blob to the Azure Blob container.
@@ -67,18 +124,14 @@ namespace Arcus.Testing
         ///     A <see cref="BlobContainerClient.Uri" /> referencing the blob container that includes the name of the account and the name of the container.
         ///     This is likely to be similar to "https://{account_name}.blob.core.windows.net/{container_name}".
         /// </param>
+        /// <param name="blobName">The name of the blob to upload.</param>
         /// <param name="blobContent">The content of the blob to upload.</param>
         /// <param name="logger">The logger to write diagnostic messages during the upload process.</param>
-        /// <exception cref="ArgumentException">Thrown when the <paramref name="blobContainerUri"/> is blank.</exception>
+        /// <exception cref="ArgumentException">Thrown when the <paramref name="blobContainerUri"/> or the <paramref name="blobName"/> is blank.</exception>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="blobContainerUri"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
-        public static async Task<TemporaryBlobFile> UploadContentAsync(Uri blobContainerUri, BinaryData blobContent, ILogger logger)
+        public static async Task<TemporaryBlobFile> UploadIfNotExistsAsync(Uri blobContainerUri, string blobName, BinaryData blobContent, ILogger logger)
         {
-            if (blobContainerUri is null)
-            {
-                throw new ArgumentNullException(nameof(blobContainerUri));
-            }
-
-            return await UploadContentAsync(blobContainerUri, blobContent, logger, configureOptions: null);
+            return await UploadIfNotExistsAsync(blobContainerUri, blobName, blobContent, logger, configureOptions: null);
         }
 
         /// <summary>
@@ -91,46 +144,59 @@ namespace Arcus.Testing
         ///     A <see cref="BlobContainerClient.Uri" /> referencing the blob container that includes the name of the account and the name of the container.
         ///     This is likely to be similar to "https://{account_name}.blob.core.windows.net/{container_name}".
         /// </param>
+        /// <param name="blobName">The name of the blob to upload.</param>
         /// <param name="blobContent">The content of the blob to upload.</param>
         /// <param name="logger">The logger to write diagnostic messages during the upload process.</param>
         /// <param name="configureOptions">The function to configure the additional options of how the blob should be uploaded.</param>
-        /// <exception cref="ArgumentException">Thrown when the <paramref name="blobContainerUri"/> is blank.</exception>
+        /// <exception cref="ArgumentException">Thrown when the <paramref name="blobContainerUri"/> or the <paramref name="blobName"/> is blank.</exception>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="blobContainerUri"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
-        public static async Task<TemporaryBlobFile> UploadContentAsync(Uri blobContainerUri, BinaryData blobContent, ILogger logger, Action<TemporaryBlobFileOptions> configureOptions)
+        public static async Task<TemporaryBlobFile> UploadIfNotExistsAsync(
+            Uri blobContainerUri,
+            string blobName,
+            BinaryData blobContent,
+            ILogger logger,
+            Action<TemporaryBlobFileOptions> configureOptions)
         {
             if (blobContainerUri is null)
             {
                 throw new ArgumentNullException(nameof(blobContainerUri));
             }
 
-            return await UploadContentAsync(new BlobContainerClient(blobContainerUri, new DefaultAzureCredential()), blobContent, logger, configureOptions);
+            var containerClient = new BlobContainerClient(blobContainerUri, new DefaultAzureCredential());
+            var blobClient = containerClient.GetBlobClient(blobName);
+
+            return await UploadIfNotExistsAsync(blobClient, blobContent, logger, configureOptions);
         }
 
         /// <summary>
         /// Uploads a temporary blob to the Azure Blob container.
         /// </summary>
-        /// <param name="containerClient">The Azure Blob container client to interact with Azure Blob storage.</param>
+        /// <param name="blobClient">The Azure Blob client to interact with Azure Blob storage.</param>
         /// <param name="blobContent">The content of the blob to upload.</param>
         /// <param name="logger">The logger to write diagnostic messages during the upload process.</param>
-        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="containerClient"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
-        public static async Task<TemporaryBlobFile> UploadContentAsync(BlobContainerClient containerClient, BinaryData blobContent, ILogger logger)
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="blobClient"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
+        public static async Task<TemporaryBlobFile> UploadIfNotExistsAsync(BlobClient blobClient, BinaryData blobContent, ILogger logger)
         {
-            return await UploadContentAsync(containerClient, blobContent, logger, configureOptions: null);
+            return await UploadIfNotExistsAsync(blobClient, blobContent, logger, configureOptions: null);
         }
 
         /// <summary>
         /// Uploads a temporary blob to the Azure Blob container.
         /// </summary>
-        /// <param name="containerClient">The Azure Blob container client to interact with Azure Blob storage.</param>
+        /// <param name="blobClient">The Azure Blob client to interact with Azure Blob storage.</param>
         /// <param name="blobContent">The content of the blob to upload.</param>
         /// <param name="logger">The logger to write diagnostic messages during the upload process.</param>
         /// <param name="configureOptions">The function to configure the additional options of how the blob should be uploaded.</param>
-        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="containerClient"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
-        public static async Task<TemporaryBlobFile> UploadContentAsync(BlobContainerClient containerClient, BinaryData blobContent, ILogger logger, Action<TemporaryBlobFileOptions> configureOptions)
+        /// <exception cref="ArgumentNullException">Thrown when the <paramref name="blobClient"/> or the <paramref name="blobContent"/> is <c>null</c>.</exception>
+        public static async Task<TemporaryBlobFile> UploadIfNotExistsAsync(
+            BlobClient blobClient,
+            BinaryData blobContent,
+            ILogger logger,
+            Action<TemporaryBlobFileOptions> configureOptions)
         {
-            if (containerClient is null)
+            if (blobClient is null)
             {
-                throw new ArgumentNullException(nameof(containerClient));
+                throw new ArgumentNullException(nameof(blobClient));
             }
 
             if (blobContent is null)
@@ -142,12 +208,35 @@ namespace Arcus.Testing
             var options = new TemporaryBlobFileOptions();
             configureOptions?.Invoke(options);
 
-            BlobClient blobClient = containerClient.GetBlobClient(options.BlobName);
+            (bool createdByUs, BinaryData originalData) = await EnsureBlobContentCreatedAsync(blobClient, blobContent, options, logger);
 
-            logger.LogDebug("Uploading Azure Blob '{BlobName}' to container '{ContainerName}'", options.BlobName, blobClient.BlobContainerName);
-            await blobClient.UploadAsync(blobContent, options.OverrideExistingBlob);
+            return new TemporaryBlobFile(blobClient, createdByUs, originalData, options, logger);
+        }
 
-            return new TemporaryBlobFile(blobClient, logger);
+        private static async Task<(bool createdByUs, BinaryData originalData)> EnsureBlobContentCreatedAsync(
+            BlobClient client, 
+            BinaryData newContent,
+            TemporaryBlobFileOptions options, 
+            ILogger logger)
+        {
+            if (await client.ExistsAsync())
+            {
+                logger.LogDebug("Azure Blob '{BlobName}' already exists in container '{ContainerName}'", client.Name, client.BlobContainerName);
+                BlobDownloadResult originalContent = await client.DownloadContentAsync();
+
+                if (options.OnSetup.Override)
+                {
+                    logger.LogDebug("Override existing Azure Blob '{BlobName}' in container '{ContainerName}'", client.Name, client.BlobContainerName);
+                    await client.UploadAsync(newContent, overwrite: true);
+                }
+
+                return (createdByUs: false, originalContent.Content);
+            }
+
+            logger.LogDebug("Uploading Azure Blob '{BlobName}' to container '{ContainerName}'", client.Name, client.BlobContainerName);
+            await client.UploadAsync(newContent);
+            
+            return (createdByUs: true, originalData: null);
         }
 
         /// <summary>
@@ -156,8 +245,17 @@ namespace Arcus.Testing
         /// <returns>A task that represents the asynchronous dispose operation.</returns>
         public async ValueTask DisposeAsync()
         {
-            _logger.LogDebug("Deleting Azure Blob '{BlobName}' from container '{ContainerName}'", _blobClient.Name, _blobClient.BlobContainerName);
-            await _blobClient.DeleteIfExistsAsync();
+            if (!_createdByUs && _originalData != null && _options.OnTeardown.Content != UponBlobDelete.DeleteIfExisted)
+            {
+                _logger.LogDebug("Reverting Azure Blob '{BlobName}' original content in container '{ContainerName}'", _blobClient.Name, _blobClient.BlobContainerName);
+                await _blobClient.UploadAsync(_originalData, overwrite: true); 
+            }
+
+            if (_createdByUs || _options.OnTeardown.Content is UponBlobDelete.DeleteIfExisted)
+            {
+                _logger.LogDebug("Deleting Azure Blob '{BlobName}' from container '{ContainerName}'", _blobClient.Name, _blobClient.BlobContainerName);
+                await _blobClient.DeleteIfExistsAsync();
+            }
         }
     }
 }
